+++ internal/public/stoptree/stoptree_test.go
@@ -12,15 +12,15 @@ type Querier struct {
 	tdb.Querier
 }
 
-func (q Querier) ListStopsInStopTree(ctx context.Context, basePk int32) ([]tdb.Stop, error) {
+func (q Querier) ListStopsInStopTree(ctx context.Context, basePk int64) ([]tdb.Stop, error) {
 	return []tdb.Stop{
 		{
 			Pk:           1,
-			ParentStopPk: sql.NullInt32{Valid: true, Int32: 3},
+			ParentStopPk: sql.NullInt64{Valid: true, Int64: 3},
 		},
 		{
 			Pk:           2,
-			ParentStopPk: sql.NullInt32{Valid: true, Int32: 3},
+			ParentStopPk: sql.NullInt64{Valid: true, Int64: 3},
 		},
 		{
 			Pk: 3,
@@ -31,17 +31,17 @@ func (q Querier) ListStopsInStopTree(ctx context.Context, basePk int32) ([]tdb.S
 func Test_Descendent(t *testing.T) {
 	stopTree, _ := NewStopTree(context.Background(), Querier{}, 3)
 	actual := stopTree.DescendentPks()
-	expected := []int32{1, 2, 3}
+	expected := []int64{1, 2, 3}
 	if !compareList(actual, expected) {
 		t.Errorf("Actual %v != expected %v", actual, expected)
 	}
 }
 
-func compareList(a, b []int32) bool {
+func compareList(a, b []int64) bool {
 	if len(a) != len(b) {
 		return false
 	}
-	aMap := map[int32]bool{}
+	aMap := map[int64]bool{}
 	for _, aElem := range a {
 		aMap[aElem] = true
 	}
